Refactoring of mpl latex settings (Cherry picked from wip:labor_umit)
