+++ pymoskito/__init__.py
@@ -10,12 +10,15 @@ mpl.use('Qt5Agg')
 os.environ["PYQTGRAPH_QT_LIB"] = "PyQt5"
 
 from .registry import *
-from .generic_processing_modules import *
-from .generic_simulation_modules import *
+
 from .processing_core import *
 from .processing_gui import *
+from .generic_processing_modules import *
+
 from .simulation_gui import *
 from .simulation_modules import *
+from .generic_simulation_modules import *
+
 from .tools import *
 from .controltools import *
 from .visualization import *
